% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderatedMediationSem2.R
\name{plot.moderatedMediationSem}
\alias{plot.moderatedMediationSem}
\title{Makes simple slop plots of moderatedMediationSem object}
\usage{
\method{plot}{moderatedMediationSem}(x, ..., digits = 3)
}
\arguments{
\item{x}{}

\item{...}{}

\item{digits}{}
}
\value{
simple slope plots for each mediator and simple slopes parameter estimates
}
\description{
Makes simple slop plots of moderatedMediationSem object
}
